+++ app/Livewire/Pages/Sheets/Index.php
@@ -2,8 +2,8 @@
 
 namespace App\Livewire\Pages\Sheets;
 
-use App\Models\Part;
 use App\Models\Sheet;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Attributes\On;
 use Livewire\Component;
 
@@ -11,6 +11,10 @@ class Index extends Component
 {
     public function deleteSheet(Sheet $sheet): void
     {
+        if (! Auth::check() || ! Auth::user()->hasVerifiedEmail()) {
+            return;
+        }
+
         $sheet->delete();
 
         $this->dispatch('sheet-deleted');
